Implement search using cache
It does not work exactly the same way as it works with eprints but I
don't think this is going to be a serious problem.

This project is not something that must ever reach production grade
quality, and I was interested to experiment with graph databases.

Plus, the main use case for the cache is to allow us to work on all the
records at the same time, for research.