Specify rules w/ Turing complete domain language
Allows to express anything at the price that now the input file
is no longer a CSV, but rather JavaScript.

The possibility of using this technique to express rules to transform
data was discussed with @giuseppefutia two weeks ago, so tagging him
in this commit since guess he could be interested.