Express compliance using float value
Anything lower than 0.5 is considered false, all other values
are considered obviously true.

This specific encoding of compliance allows to express:

a) which specific clause was matched (this is used for example
   in `date_made_open` compliance computation)

b) cases where it's more yes then no and viceversa

The `is_compliant` field is unchanged (this was verified by
checking the result of `node index.js -t` against the previous
result emitted by that command). A new field `is_compliant_float`
has been added to allow one to access the nuanced value.

This more or less supersedes #10, but the part of #10 emitting
the proper strings could still be useful.