Add lib/index.js: simplified express based server
By default it dumps the current roarmap database as JSON in the
current working directory as `roarmap-dump.json`. To this end, it
uses jsdom to scrape roarmap's website.

When invoked with `-l`, provides the same JSON via an API.

The implementation of this new server/scraper obsoletes the
scraper implemented using phantomjs, which is removed accordingly.

While there, .gitignore annoying .DS_Store files.