server.js: deal with unhandled errors w/ cluster
In case some piece of code throws an exception that we did not
anticipate, be prepared by running most of the code in a background
process and respawning such process in case of its death.

After reading about catching any exception and about domains, this
really seems to me the optimal way to deal with the unexpected.